requirements.txt
# pip install --upgrade pip
# pip install -r requirements.txt
langchain==0.3.24
langchain-core==0.3.55
langchain-community==0.3.22
langchain-huggingface==0.1.2
langchain-groq==0.3.2
huggingface==0.0.1
huggingface-hub==0.30.2
python-dotenv==1.1.0
pypdf==5.4.0
faiss-cpu==1.10.0
sentence-transformers==4.1.0
streamlit==1.43.2